chore: renaming directories
