+++ apps/web/app/friendship/friends/components/chat/chat-topbar.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Avatar, AvatarImage } from '@/components/ui/avatar'
-import { UserData } from '@/app/friendship/chat/data';
+import { UserData } from '@/app/friendship/friends/data';
 import { Info, Phone, Video } from 'lucide-react';
 import Link from 'next/link';
 import { cn } from '@/lib/utils';
