+++ apps/web/app/friendship/friends/components/emoji-picker.tsx
